pyproject.toml
[tool.poetry]
name = "course-progect-1"
version = "0.1.0"
description = ""
authors = ["Nafikova_Julia <nafikova91@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
load-dotenv = "^0.1.0"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "/.git"

[tool.isort]
line-length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"
